feat: populate QueryResult.fields with names and types
I found myself missing this in handling query results, and felt the simplest
way to phrase the feature request was to write a test. I want to read both
column names and types, for the query result.

Included is a draft implementation in `SelectExec` that satisfies this one
test. But I notice that there is corresponding column naming logic in
`src/transforms/selection.ts` with more involved name defaulting based on the
column expressions, so I suspect probably that should be flowing through to the
query result instead.